configurator: replace to C.
